tests/wpt/web-platform-tests/webxr/xrDevice_requestSession_immersive_no_gesture.https.html

<!DOCTYPE html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script src="resources/webxr_util.js"></script>
  <script src="resources/webxr_test_constants.js"></script>
  <script>
    xr_promise_test(
      "Requesting immersive session outside of a user gesture rejects",
      (t) => {
        return navigator.xr.test.simulateDeviceConnection(TRACKED_IMMERSIVE_DEVICE)
          .then( (controller) => promise_rejects_dom(
            t, 'SecurityError', navigator.xr.requestSession('immersive-vr')));
      });
  </script>
</body>
